requirements.txt
pandas
wordcloud
matplotlib
base64
python-dateutil
taipy
taipy-gui

